requirements.txt
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
django-cors-headers==2.4.0
django-filter==2.0.0
django-oauth-toolkit==1.2.0
djangorestframework==3.9.0
drf-nested-routers==0.91
drf-yasg==1.16.0
idna==2.8
importlib-metadata==0.18
inflection==0.3.1
itypes==1.1.0
MarkupSafe==1.1.1
more-itertools==7.1.0
oauthlib==3.0.2
pluggy==0.12.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
py==1.8.0
pytest==4.0.2
pytest-django==3.4.4
python-dotenv==0.10.1
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
uritemplate==3.0.0
